Finally done with chapter 5

+++ src/SDLGameObject.cpp
@@ -14,13 +14,13 @@ SDLGameObject::SDLGameObject(const LoaderParams *pParams) : GameObject(pParams),
 
 void SDLGameObject::draw(){
 	if(m_velocity.getX() > 0){
-		TextureManager::Instance()->drawFrame(m_textureID, (Uint32)m_position.getX(),
-			(Uint32)m_position.getY(), m_width, m_height, m_currentRow, m_currentFrame,
+		TextureManager::Instance()->drawFrame(m_textureID, m_position.getX(),
+			m_position.getY(), m_width, m_height, m_currentRow, m_currentFrame,
 			TheGame::Instance()->getRenderer(), SDL_FLIP_HORIZONTAL);
 	}
 	else{
-		TextureManager::Instance()->drawFrame(m_textureID, (Uint32)m_position.getX(), 
-			(Uint32)m_position.getY(), m_width, m_height, m_currentRow, m_currentFrame,
+		TextureManager::Instance()->drawFrame(m_textureID, m_position.getX(), 
+			m_position.getY(), m_width, m_height, m_currentRow, m_currentFrame,
 			TheGame::Instance()->getRenderer());
 	}
 }
